Refactors project structure to remove pkg/ and move commands to internal/
The idea here is to have a logical separation between code that can be used as a library and code which is for CLI specific operations. This enforces packages outside of internal/ to have a very friendly library interface and will allow others to easily include them as a dependency without inheriting unecessary logic.